// src/api/models/ApiBlockGridModel.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Umbraco Delivery API
 * You can find out more about the Umbraco Delivery API in [the documentation](https://docs.umbraco.com/umbraco-cms/reference/content-delivery-api).
 *
 * The version of the OpenAPI document: Latest
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { ApiBlockGridItemModel } from './ApiBlockGridItemModel';
import {
    ApiBlockGridItemModelFromJSON,
    ApiBlockGridItemModelFromJSONTyped,
    ApiBlockGridItemModelToJSON,
} from './ApiBlockGridItemModel';

/**
 * 
 * @export
 * @interface ApiBlockGridModel
 */
export interface ApiBlockGridModel {
    /**
     * 
     * @type {number}
     * @memberof ApiBlockGridModel
     */
    gridColumns?: number;
    /**
     * 
     * @type {Array<ApiBlockGridItemModel>}
     * @memberof ApiBlockGridModel
     */
    items?: Array<ApiBlockGridItemModel>;
}

/**
 * Check if a given object implements the ApiBlockGridModel interface.
 */
export function instanceOfApiBlockGridModel(value: object): boolean {
    let isInstance = true;

    return isInstance;
}

export function ApiBlockGridModelFromJSON(json: any): ApiBlockGridModel {
    return ApiBlockGridModelFromJSONTyped(json, false);
}

export function ApiBlockGridModelFromJSONTyped(json: any, ignoreDiscriminator: boolean): ApiBlockGridModel {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'gridColumns': !exists(json, 'gridColumns') ? undefined : json['gridColumns'],
        'items': !exists(json, 'items') ? undefined : ((json['items'] as Array<any>).map(ApiBlockGridItemModelFromJSON)),
    };
}

export function ApiBlockGridModelToJSON(value?: ApiBlockGridModel | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'gridColumns': value.gridColumns,
        'items': value.items === undefined ? undefined : ((value.items as Array<any>).map(ApiBlockGridItemModelToJSON)),
    };
}