// src/api/models/ImageWithCaptionElementModel.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { IApiElementModelBase } from './IApiElementModelBase';
import type { ImageWithCaptionPropertiesModel } from './ImageWithCaptionPropertiesModel';
export type ImageWithCaptionElementModel = (IApiElementModelBase & {
    properties?: ImageWithCaptionPropertiesModel;
});