// src/api/models/RichTextElementModel.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { IApiElementModelBase } from './IApiElementModelBase';
import type { RichTextPropertiesModel } from './RichTextPropertiesModel';
export type RichTextElementModel = (IApiElementModelBase & {
    properties?: RichTextPropertiesModel;
});